🚀 add start and build scripts to the package.json
