+++ src/back-end/package.json
@@ -5,8 +5,10 @@
   "main": "index.js",
   "private": "true",
   "scripts": {
+    "start":"node build/app.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon src/app.ts"
+    "dev": "nodemon src/app.ts",
+    "build": "tsc -p ."
   },
   "keywords": [],
   "author": "",
